Avoid calling negfix8 concurrently because it doesn't work
A very helpful user (Christopher Meyer) discovered that running this
script on a folder multiple times led to slightly different results,
and then narrowed it down to the fact that Negfix8 is only storing a
single temporary profile for each run, which means  that running
multiple instances at once would cause it to incorrectly use that
temporary profile for multiple/wrong images.  As this is an issue with
Negfix8 (which could be solved by keying the temporary profile data
based on file location/name), for now we’ll simply have to avoid
running Negfix8 concurrently, which is going to make things slower on
most modern CPUs.

Negfix8 does let you save this profile data to a specific name and opt
to reuse it later, but I don’t see an obvious way that it would let us
simply specify a unique name for a single pass.

GH Issue:
https://github.com/alibosworth/pakon-planar-raw-converter/issues/6