Switch to saving initial pre-inversion TIFF with interleaved pixel order
Somehow, despite identical visual results, when the per-channel
pixel-order tiff is run through NegFix8, it produces a file that has a
choppy/toothy histogram in Photoshop.  It was per-channel because of
the initial `-interlace plane` flag to parse it as planar, but needs to
be then set to -`interlace none`.

Thanks to Trevor Crump on the Pakon Facebook group for bringing this to
my attention!
https://www.facebook.com/groups/PakonF135/permalink/1701517296533967/